[feat] dm, dm-message entity, domain files
