+++ src/feature/api/notification/infrastructure/dm.entity.ts
@@ -1,11 +1,15 @@
 import { DateTimeType, Entity, PrimaryKey, Property } from '@mikro-orm/core';
+import { v4 } from 'uuid';
 
 @Entity({ tableName: 'dm' })
 export class DmEntity {
-  @PrimaryKey({ type: 'uuid' })
+  @PrimaryKey({ type: 'uuid', defaultRaw: 'gen_random_uuid()' })
+  id: string = v4();
+
+  @Property({ type: 'uuid' })
   user1Id: string;
 
-  @PrimaryKey({ type: 'uuid' })
+  @Property({ type: 'uuid' })
   user2Id: string;
 
   @Property({ type: DateTimeType, defaultRaw: 'current_timestamp' })
